requirements.txt
Flask
Flask-JWT-Extended
pymysql
cryptography
python-dotenv